♻️ Refactor: Extract restaurant menu API logic into custom hook (useRestaurantMenu)
